requirements.txt
streamlit>=1.28.0
pyrebase4>=4.7.1
pandas>=2.0.0
plotly>=5.17.0

